+++ EmprestimoApi/Emprestimo.Business/Models/BoletoEmprestimo.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int NumeroParcela { get; set; }
-        public double ValorDaParcela { get; set; }
+        public decimal ValorDaParcela { get; set; }
         public DateTime DataDePagamento { get; set; }
 
         public int EmprestimoId { get; set; }
